+++ gerbera-web/test/client/gerbera.items.spec.js
@@ -335,7 +335,7 @@ describe('Gerbera Items', () => {
 
       Items.loadEditItem(editItemResponse);
 
-      expect(editObjectType.val()).toBe('item');
+      expect(editObjectType.val()).toBe('video');
       expect(editTitle.val()).toBe('Test.mp4');
       expect(editLocation.val()).toBe('/folder/location/Test.mp4');
       expect(editClass.val()).toBe('object.item.videoItem');